[INLONG-2045] sort-sdk support Prometheus metrics
[INLONG-2045] sort-sdk support Prometheus metrics

+++ inlong-sdk/sort-sdk/src/main/java/org/apache/inlong/sdk/sort/api/MetricReporter.java
@@ -19,7 +19,7 @@ package org.apache.inlong.sdk.sort.api;
 
 public interface MetricReporter {
 
-    boolean report(String monitorName, String[] keys, double[] values);
+    boolean report(String monitorName, String[] keys, long[] values);
 
     void close();
 }
